Added standalone, changed to non-client class
Changed NodeCrawler to a library class. Added a standalone client that
uses NodeCrawler. Added TODOs for TUI and CLI options
